--- a/Library/Homebrew/abstract_command.rb
+++ b/Library/Homebrew/abstract_command.rb
@@ -24,7 +24,13 @@ module Homebrew
       attr_reader :args_class
 
       sig { returns(String) }
-      def command_name = Utils.underscore(T.must(name).split("::").fetch(-1)).tr("_", "-").delete_suffix("-cmd")
+      def command_name
+        require "utils"
+
+        Utils.underscore(T.must(name).split("::").fetch(-1))
+             .tr("_", "-")
+             .delete_suffix("-cmd")
+      end
 
       # @return the AbstractCommand subclass associated with the brew CLI command name.
       sig { params(name: String).returns(T.nilable(T.class_of(AbstractCommand))) }