--- a/Library/Homebrew/api.rb
+++ b/Library/Homebrew/api.rb
@@ -8,7 +8,6 @@ require "warnings"
 Warnings.ignore :default_gems do
   require "base64" # TODO: Add this to the Gemfile or remove it before moving to Ruby 3.4.
 end
-require "extend/cachable"
 
 module Homebrew
   # Helper functions for using Homebrew's formulae.brew.sh API.