--- a/Library/Homebrew/cmd/tap-info.rb
+++ b/Library/Homebrew/cmd/tap-info.rb
@@ -23,6 +23,8 @@ module Homebrew
 
       sig { override.void }
       def run
+        require "tap"
+
         taps = if args.installed?
           Tap
         else