--- a/Library/Homebrew/formulary.rb
+++ b/Library/Homebrew/formulary.rb
@@ -4,12 +4,14 @@
 require "digest/sha2"
 require "extend/cachable"
 require "tab"
+require "utils"
 require "utils/bottles"
 require "service"
 require "utils/curl"
 require "deprecate_disable"
 require "extend/hash/deep_transform_values"
 require "extend/hash/keys"
+require "tap"
 
 # The {Formulary} is responsible for creating instances of {Formula}.
 # It is not meant to be used directly from formulae.