--- a/Library/Homebrew/keg_relocate.rb
+++ b/Library/Homebrew/keg_relocate.rb
@@ -271,6 +271,8 @@ class Keg
         next true if pn.directory?
         next false if pn.basename.to_s == "orig-prefix.txt" # for python virtualenvs
         next true if pn == self/".brew/#{name}.rb"
+
+        require "metafiles"
         next true if Metafiles::EXTENSIONS.include?(pn.extname)
 
         if pn.text_executable?