--- a/Library/Homebrew/startup/ruby_path.rb
+++ b/Library/Homebrew/startup/ruby_path.rb
@@ -1,7 +1,5 @@
 # typed: true
 # frozen_string_literal: true
 
-require "rbconfig"
-
 RUBY_PATH = Pathname.new(RbConfig.ruby).freeze
 RUBY_BIN = RUBY_PATH.dirname.freeze