--- a/Library/Homebrew/utils.rb
+++ b/Library/Homebrew/utils.rb
@@ -1,27 +1,7 @@
 # typed: true
 # frozen_string_literal: true
 
-require "time"
-
-require "utils/analytics"
-require "utils/backtrace"
-require "utils/curl"
-require "utils/fork"
-require "utils/formatter"
-require "utils/gems"
-require "utils/git"
-require "utils/git_repository"
-require "utils/github"
-require "utils/gzip"
-require "utils/inreplace"
-require "utils/link"
-require "utils/popen"
-require "utils/repology"
-require "utils/svn"
-require "utils/tty"
-require "tap_constants"
-require "PATH"
-require "extend/kernel"
+require "context"
 
 module Homebrew
   extend Context
@@ -62,6 +42,8 @@ module Homebrew
 
         method = instance_method(name)
         define_method(name) do |*args, &block|
+          require "time"
+
           time = Time.now
 
           begin