… Library/Homebrew/cask/audit.rb
… Library/Homebrew/formula.rb