@samford 2024-07-14
> This PR broke `brew bump`

@yurikoles Thanks for reporting this. I was able to reproduce the error locally and it should be fixed by #17717.